Split YANG getter-at-path out to own module
We'll also start moving utilities from lib.ptree.ptree into
lib.yang.path_data.
